Apparently not: We need to do a merge function
There doesn't seem to be a way (via destructuring, elsewise)
to take environment parameters from webpack (which send an object
literal) and have them merge with a default.

It's a wholesale override of the object if there's a default. It's not
individual keys being looked at filling in what's missing.